requirements.txt
requests
browser-cookie3
Pillow
argparse
